Split and add space pages
