+++ front-spa/src/app/App.tsx
@@ -22,6 +22,25 @@ import { PaymentProcessingPage } from "@dust-tt/front/components/pages/workspace
 import { SubscriptionPage } from "@dust-tt/front/components/pages/workspace/subscription/SubscriptionPage";
 import { WorkspaceSettingsPage } from "@dust-tt/front/components/pages/workspace/WorkspaceSettingsPage";
 
+// Space pages
+import { DataSourceViewPage } from "@dust-tt/front/components/pages/spaces/DataSourceViewPage";
+import { SpaceActionsPage } from "@dust-tt/front/components/pages/spaces/SpaceActionsPage";
+import { SpaceAppsListPage } from "@dust-tt/front/components/pages/spaces/SpaceAppsListPage";
+import { SpaceCategoryPage } from "@dust-tt/front/components/pages/spaces/SpaceCategoryPage";
+import { SpacePage } from "@dust-tt/front/components/pages/spaces/SpacePage";
+import { SpacesRedirectPage } from "@dust-tt/front/components/pages/spaces/SpacesRedirectPage";
+import { SpaceTriggersPage } from "@dust-tt/front/components/pages/spaces/SpaceTriggersPage";
+
+// App pages
+import { AppSettingsPage } from "@dust-tt/front/components/pages/spaces/apps/AppSettingsPage";
+import { AppSpecificationPage } from "@dust-tt/front/components/pages/spaces/apps/AppSpecificationPage";
+import { AppViewPage } from "@dust-tt/front/components/pages/spaces/apps/AppViewPage";
+import { DatasetPage } from "@dust-tt/front/components/pages/spaces/apps/DatasetPage";
+import { DatasetsPage } from "@dust-tt/front/components/pages/spaces/apps/DatasetsPage";
+import { NewDatasetPage } from "@dust-tt/front/components/pages/spaces/apps/NewDatasetPage";
+import { RunPage } from "@dust-tt/front/components/pages/spaces/apps/RunPage";
+import { RunsPage } from "@dust-tt/front/components/pages/spaces/apps/RunsPage";
+
 const router = createBrowserRouter(
   [
     {
@@ -62,6 +81,28 @@ const router = createBrowserRouter(
           path: "labs/mcp_actions/:agentId",
           element: <AgentMCPActionsPage />,
         },
+
+        // Spaces
+        { path: "spaces", element: <SpacesRedirectPage /> },
+        { path: "spaces/:spaceId", element: <SpacePage /> },
+        { path: "spaces/:spaceId/categories/actions", element: <SpaceActionsPage /> },
+        { path: "spaces/:spaceId/categories/apps", element: <SpaceAppsListPage /> },
+        { path: "spaces/:spaceId/categories/triggers", element: <SpaceTriggersPage /> },
+        { path: "spaces/:spaceId/categories/:category", element: <SpaceCategoryPage /> },
+        {
+          path: "spaces/:spaceId/categories/:category/data_source_views/:dataSourceViewId",
+          element: <DataSourceViewPage />,
+        },
+
+        // Apps
+        { path: "spaces/:spaceId/apps/:aId", element: <AppViewPage /> },
+        { path: "spaces/:spaceId/apps/:aId/settings", element: <AppSettingsPage /> },
+        { path: "spaces/:spaceId/apps/:aId/specification", element: <AppSpecificationPage /> },
+        { path: "spaces/:spaceId/apps/:aId/datasets", element: <DatasetsPage /> },
+        { path: "spaces/:spaceId/apps/:aId/datasets/new", element: <NewDatasetPage /> },
+        { path: "spaces/:spaceId/apps/:aId/datasets/:name", element: <DatasetPage /> },
+        { path: "spaces/:spaceId/apps/:aId/runs", element: <RunsPage /> },
+        { path: "spaces/:spaceId/apps/:aId/runs/:runId", element: <RunPage /> },
       ],
     },
     { path: "*", element: <Navigate to="/w/DevWkSpace/me" replace /> },
